requirements.txt
Flask==0.10.1
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
itsdangerous==0.24
pyaml==14.05.7
